requirements.txt
python-telegram-bot[job-queue]==20.3
playwright==1.40.0
python-dotenv==1.0.0
